Use FFT for MAC projection cases (#152)
For level 0 MAC projection with all periodic boundaries or periodic in
the first two dimensions and Neumann in the last dimension, there is now
the option of using FFT. To compile with FFT support, use `USE_FFT=TRUE`
for GNU Make and `HYDRO_FFT=ON` (which will turn on AMReX_FFT) for
CMake. At runtime, one needs to specify `mac_proj.use_fft = 1`. If the
conditions for FFT support are not satisfied, a runtime error may occur.
So don't set `mac_proj.use_fft=1` unless FFT is supported for the
problem.

Note that the current implementation is still experimental.