Simplify cache cleanup actions (#149)
In our old approach, a workflow file contains a job that uploads the PR
number as an artifact. While the PR is still open, the workflow_run
triggered by it will download the artifact and use the information to
clean up all except the last used cache associated with that original
workflow. When a PR is merged or closed, there will be a post-pr
workflow that uploads the PR number as an artifact and triggers a
workflow_run that clean up all caches associated with the PR. The reason
we did it this way was in the cache cleanup workflows, we did not find
an easy way to get the number of the PR triggering them. This is not
convenient because we have to add jobs uploading artifacts to workflow
files.

After some experiments, we have found a reliable way to find the PR
number without using artifacts. The workflow_run's payload always
contains the head SHA of the commit that triggers it, whether the PR
comes from a fork or not. We can then use `gh pr list` to search for
that head and obtain the PR number.