Allow outflow to be zero when inflow is zero while enforcing solvability (#159)
This fixes a bug in the existing code which requires the outflow to be
non-zero even if the inflow is zero for the direction-dependent
boundaries. It should be okay for the outflow to be zero when inflow is
zero, e.g. for finer levels that are not touching the boundary at all,
or even in general.