Allow Hydro to use its own EB setting instead of inheriting from AMReX (#158)
This allows compiling Hydro with `HYDRO_EB=OFF` even if AMReX is
compiled with `AMReX_EB=ON`. This was required to compile AMR-Wind
(which requires EB=OFF in Hydro) and ERF (which now requires EB=ON in
AMReX) for the ERF/AMR-Wind coupling driver.

Note: do we need to implement this for the GNUmake build system too?