Fix multiple-grid bug by including non-corner ghost cells in outflow-correction (#161)
This PR fixes https://github.com/Exawind/amr-wind/issues/1496, an issue
in the inflow-outflow solvability functions.

Each grid did not see the outflow-correction velocity update for
non-corner boundary cells in their ghost region as the responsibility
for updating these was with the neighboring grid. A FillBoundary does
not fill these cells. This caused the nodal projection to not converge
as discussed in https://github.com/Exawind/amr-wind/issues/1496.

This change fixes this issue by including these non-corner ghost cells
of a grid during the outflow-correction routine.